refactor: separate config validation
