+++ src/sparsevllm/configs/common.py
@@ -1,0 +1,57 @@
+"""Shared coercion helpers for configuration modules."""
+
+import os
+from typing import Any
+
+def _coerce_bool_config(name: str, value: Any) -> bool:
+    if isinstance(value, bool):
+        return value
+    if isinstance(value, str):
+        raw = value.strip().lower()
+        if raw in {"1", "true", "yes", "on"}:
+            return True
+        if raw in {"0", "false", "no", "off"}:
+            return False
+    raise ValueError(f"{name} must be a boolean or explicit true/false string, got {value!r}.")
+
+
+def _coerce_optional_positive_int(name: str, value: Any) -> int | None:
+    if value is None:
+        return None
+    if isinstance(value, bool):
+        raise ValueError(f"{name} must be a positive integer, got {value!r}.")
+    if isinstance(value, int):
+        parsed = value
+    elif isinstance(value, str):
+        raw = value.strip()
+        if not raw.isdecimal():
+            raise ValueError(f"{name} must be a positive integer, got {value!r}.")
+        parsed = int(raw)
+    else:
+        raise ValueError(f"{name} must be a positive integer, got {value!r}.")
+    if parsed <= 0:
+        raise ValueError(f"{name} must be > 0 when set, got {parsed}.")
+    return parsed
+
+
+def _resolve_long_prefill_offload_threshold(configured: Any) -> int:
+    raw = os.getenv("SPARSEVLLM_LONG_PREFILL_OFFLOAD_MIN_TOKENS")
+    legacy_raw = os.getenv("SPARSEVLLM_DEFERRED_PREFILL_MIN_TOKENS")
+    if raw is not None and legacy_raw is not None and raw != legacy_raw:
+        raise ValueError(
+            "SPARSEVLLM_LONG_PREFILL_OFFLOAD_MIN_TOKENS and "
+            "SPARSEVLLM_DEFERRED_PREFILL_MIN_TOKENS are both set with different values."
+        )
+    value = raw if raw is not None else legacy_raw
+    resolved = _coerce_optional_positive_int(
+        "long_prefill_offload_threshold",
+        configured if value is None else value,
+    )
+    if resolved is None:
+        raise ValueError("long_prefill_offload_threshold must be a positive integer.")
+    return int(resolved)
+
+
+
+def _model_path_basename(model_path: str) -> str:
+    return str(model_path).rstrip("/").split("/")[-1]
