+++ tests/test_config_groups.py
@@ -1,3 +1,6 @@
+import ast
+import inspect
+import textwrap
 from dataclasses import fields
 
 from sparsevllm.config import Config as PublicConfig
@@ -39,3 +42,30 @@ def test_config_composes_disjoint_responsibility_groups():
 def test_config_groups_are_keyword_only():
     for group in CONFIG_GROUPS:
         assert all(item.kw_only for item in fields(group) if item.init)
+
+
+def test_runtime_constructor_only_orchestrates_config_stages():
+    tree = ast.parse(textwrap.dedent(inspect.getsource(Config.__post_init__)))
+    forbidden = (ast.For, ast.If, ast.Raise, ast.Try, ast.While)
+
+    assert not any(isinstance(node, forbidden) for node in ast.walk(tree))
+
+    calls = {
+        node.func.id
+        for node in ast.walk(tree)
+        if isinstance(node, ast.Call) and isinstance(node.func, ast.Name)
+    }
+    assert {
+        "normalize_bootstrap",
+        "normalize_sparse_method_name",
+        "normalize_prefix_cache",
+        "normalize_scheduling",
+        "normalize_deltakv_storage",
+        "normalize_platform",
+        "normalize_decode_cuda_graph",
+        "load_and_validate_model",
+        "normalize_sparse_methods",
+        "finalize_prefix_cache",
+        "validate_deltakv_runtime",
+        "finalize_sparse_layout",
+    } <= calls
